@@ -11,11 +11,11 @@
   },
   "scripts": {
     "start": "NODE_PATH=. nodemon server/server.js --exec babel-node -c config/nodemon.json",
-    "test": "mocha --compilers js:babel-core/register --require ./config/mocha.js server/**/*.spec.js",
-    "test:watch": "npm run test -- --watch",
-    "coverage": "istanbul cover _mocha -- --compilers js:babel-core/register --require ./config/test.js server/**/*.spec.js --report lcovonly",
-    "lint": "eslint server client config test -c config/eslint.js",
-    "fix": "npm run lint -- --fix",
+    "watch": "npm-watch",
+    "lint": "eslint --fix -c config/eslint.js server client config test",
+    "stylelint": "stylelint client/**/*.js",
+    "test": "jest server client",
+    "coverage": "npm run test -- --coverage",
     "e2e:setup": "node_modules/selenium-standalone/bin/selenium-standalone install",
     "e2e": "node_modules/.bin/nightwatch --config config/nightwatch.js",
     "build:server": "babel server -d dist/server --copy-files --ignore server/**/*.spec.js",
@@ -24,8 +24,12 @@
     "serve": "NODE_PATH=./dist NODE_ENV=production node dist/server/server.js",
     "serve:fresh": "npm run build && npm run serve"
   },
+  "watch": {
+    "lint": "server client config test",
+    "stylelint": "client/**/*.js",
+    "test": "client server"
+  },
   "devDependencies": {
-    "autoprefixer": "^6.5.1",
     "babel-cli": "^6.16.0",
     "babel-core": "^6.18.2",
     "babel-eslint": "^7.0.0",
@@ -42,40 +46,38 @@
     "babel-preset-react-hmre": "^1.1.1",
     "babel-register": "^6.16.3",
     "babel-runtime": "^6.9.2",
-    "chai": "^3.5.0",
-    "chai-immutable": "^1.6.0",
     "chromedriver": "2.29.0",
-    "css-loader": "^0.25.0",
     "cucumber": "^2.0.0-rc.9",
     "enzyme": "^2.6.0",
-    "eslint": "^3.8.1",
-    "eslint-plugin-mocha": "^4.7.0",
-    "eslint-plugin-react": "^6.5.0",
+    "enzyme-to-json": "^1.5.1",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-jest": "^20.0.3",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-watch": "^3.1.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.10.1",
     "ignore-styles": "^5.0.1",
     "image-webpack-loader": "^3.0.0",
     "istanbul": "^1.0.0-alpha.2",
-    "jsdom": "^9.8.3",
+    "jest": "^20.0.4",
     "json-loader": "^0.5.4",
-    "mocha": "^3.1.2",
     "morgan": "^1.8.1",
     "nightwatch": "^0.9.9",
-    "node-sass": "^4.1.1",
     "nodemon": "^1.11.0",
-    "postcss-loader": "^1.0.0",
-    "postcss-smart-import": "^0.6.6",
-    "precss": "^1.4.0",
+    "npm-watch": "^0.1.9",
     "react-addons-perf": "^15.3.2",
     "react-addons-test-utils": "^15.4.1",
     "react-hot-loader": "^3.0.0-beta.6",
     "react-transform-hmr": "^1.0.4",
-    "redux-devtools-extension": "^1.0.0",
-    "sass-loader": "^4.0.2",
+    "redux-devtools-extension": "^2.13.2",
     "selenium-standalone": "^5.8.0",
     "selenium-webdriver": "3.4.0",
-    "sinon": "^1.17.6",
-    "style-loader": "^0.13.1",
+    "stylelint": "^7.10.1",
+    "stylelint-config-standard": "^16.0.0",
+    "stylelint-processor-styled-components": "^0.1.0",
     "webpack": "^1.13.1",
     "webpack-dev-middleware": "^1.8.4",
     "webpack-hot-middleware": "^2.13.1"
@@ -85,6 +87,9 @@
     "compression": "^1.6.2",
     "cors": "^2.8.3",
     "express": "^4.14.0",
+    "express-session": "^1.15.3",
+    "express-validator": "^3.2.0",
+    "favicon": "^0.0.2",
     "helmet": "^3.5.0",
     "immutable": "^3.8.1",
     "jsonwebtoken": "^7.3.0",
