@@ -1,45 +1,42 @@
 require('babel-core/register');
 
 module.exports = {
-	"src_folders": ["test"],
-	"output_folder": "reports",
-	"custom_commands_path": "",
-	"custom_assertions_path": "",
-	"page_objects_path": "",
-	"globals_path": "",
-
-	"selenium": {
-		"start_process": true,
-		"server_path": "./node_modules/selenium-standalone/.selenium/selenium-server/2.53.1-server.jar",
-		"log_path": "./reports",
-		"host": "127.0.0.1",
-		"port": 4444,
-		"cli_args": {
-			"webdriver.chrome.driver": "./node_modules/selenium-standalone/.selenium/chromedriver/2.25-x64-chromedriver",
-			"webdriver.gecko.driver": "",
-			"webdriver.edge.driver": ""
-		}
-	},
-
-	"test_settings": {
-
-		"default": {
-			"launch_url": "http://localhost",
-			"selenium_port": 4444,
-			"selenium_host": "localhost",
-			"silent": true,
-			"screenshots": {
-				"enabled": false,
-				"path": ""
-			},
-			"globals": {
-				"waitForConditionTimeout": 5000
-			},
-			"desiredCapabilities": {
-				"browserName": "chrome",
-				"javascriptEnabled": true,
-				"acceptSslCerts": true
-			}
-		}
-	}
+  src_folders: ['test'],
+  output_folder: 'reports',
+  custom_commands_path: '',
+  custom_assertions_path: '',
+  page_objects_path: '',
+  globals_path: '',
+  selenium: {
+    start_process: true,
+    server_path: './node_modules/selenium-standalone/.selenium/selenium-server/2.53.1-server.jar',
+    log_path: './reports',
+    host: '127.0.0.1',
+    port: 4444,
+    cli_args: {
+      'webdriver.chrome.driver': './node_modules/selenium-standalone/.selenium/chromedriver/2.25-x64-chromedriver',
+      'webdriver.gecko.driver': '',
+      'webdriver.edge.driver': '',
+    },
+  },
+  test_settings: {
+    default: {
+      launch_url: 'http://localhost',
+      selenium_port: 4444,
+      selenium_host: 'localhost',
+      silent: true,
+      screenshots: {
+        enabled: false,
+        path: '',
+      },
+      globals: {
+        waitForConditionTimeout: 5000,
+      },
+      desiredCapabilities: {
+        browserName: 'chrome',
+        javascriptEnabled: true,
+        acceptSslCerts: true,
+      },
+    },
+  },
 };
